+++ zukit/traits/admin-menu.php
@@ -8,13 +8,22 @@ trait zukit_AdminMenu {
 	private static $menu_split_id = 'options-privacy.php';
 	private static $menu_split_index = null;
 	private static $menu_moved_index = null;
+	private static $menu_inline_style = false;
+	private static $menu_debug = false;
+	private static $menu_disable = false;
+	private static $menu_configs = 0;
+
+	private static $detailed_log = false;
+	private static $if_log = false;
+	private static $if_class = null; //'zu_Plus';
 
 	private $shift_gap = 20;
 	protected $total_index_break = 220;		// max items count in submenu?
 
 	public function admin_menu_config() {
 		add_filter('custom_menu_order', [$this, 'admin_menu_modify']);
-		if($this->is_origin()) {
+		self::$menu_configs += 1;
+		if(!self::$menu_inline_style) {
 			$this->snippets('add_admin_inline_style',
 				'.wp-core-ui .wp-submenu .wp-menu-separator a',
 				'border-top: 1px solid;
@@ -24,6 +33,7 @@ trait zukit_AdminMenu {
 				 pointer-events: none;
 			     cursor: default;'
 			);
+			self::$menu_inline_style = true;
 		}
 	}
 
@@ -40,7 +50,14 @@ trait zukit_AdminMenu {
 
 	protected function custom_admin_menu() { return []; }
 	protected function custom_admin_submenu() { return []; }
-	protected function custom_menu_debug() { return false; }
+
+	public function toggle_menu_debug($toggle) {
+		self::$menu_debug = $toggle;
+	}
+
+	public function toggle_menu_disable($toggle) {
+		self::$menu_disable = $toggle;
+	}
 
 	// Admin menu modify ------------------------------------------------------]
 
@@ -48,12 +65,10 @@ trait zukit_AdminMenu {
 		global $submenu;
 		if(empty(self::$menu_split_index)) {
 			if(!isset($submenu[self::$default_menu_id])) return null;
-
 			$submenu_items = count($submenu[self::$default_menu_id]);
-			self::$menu_split_index = $this->get_submenu_index(self::$menu_split_id);
+			self::$menu_split_index = $this->get_menu_index(true, self::$menu_split_id);
 			$split_position = array_search(self::$menu_split_index, array_keys($submenu[self::$default_menu_id]));
 			$moved_items = $submenu_items - $split_position - 1;
-
 			// clean positions from $menu_split_index + 1
 			$this->submenu_move(
 				self::$menu_split_index + 1,
@@ -75,6 +90,15 @@ trait zukit_AdminMenu {
 	public function admin_menu_modify($menu_order) {
 	    global $menu, $submenu;
 
+		if(self::$menu_disable) return $menu_order;
+
+		// some internal debug logging
+		self::$if_log = self::$if_class === static::class;
+		if(empty(self::$menu_split_index) && self::$menu_debug) {
+			// output origin submenu order for debug purpose
+			$this->debug_print(false, false, true);
+		}
+
 		if($this->get_split_index() === null) return $menu_order;
 
 		// modify main menu items
@@ -82,21 +106,24 @@ trait zukit_AdminMenu {
 
 		if(!empty($modify)) {
 			if(isset($modify['reorder'])) {
+				$prev_action = false;
 			    foreach($modify['reorder'] as $menu_item) {
-			    	$this->menu_reorder($menu_item['menu'], $this->get_new_index($menu_item));
+					// if the item only for 'onfail' and the previous result is successful, then skip it
+					if(($menu_item['onfail'] ?? false) && $prev_action) continue;
+			    	$prev_action = $this->menu_reorder($menu_item, false);
 			    }
 			}
 
 	    	if(isset($modify['rename'])) {
 			    foreach($modify['rename'] as $menu_item) {
-			    	$index = $this->get_menu_index($menu_item['menu']);
+			    	$index = $this->get_menu_index(false, $menu_item['menu']);
 			    	if($index > 0) $menu[$index][0] = $menu_item['new_name'];
 			    }
 			}
 
 	    	if(isset($modify['remove'])) {
 			    foreach($modify['remove'] as $menu_item) {
-			    	$index = $this->get_menu_index($menu_item['menu']);
+			    	$index = $this->get_menu_index(false, $menu_item['menu']);
 			    	if($index > 0) {
 				    	unset($menu[$index]);
 						ksort($menu);
@@ -112,16 +139,18 @@ trait zukit_AdminMenu {
 			$default_id = self::$default_menu_id;
 
 			if(isset($modify['reorder'])) {
+				$prev_action = false;
 			    foreach($modify['reorder'] as $menu_item) {
-			    	$parent = $menu_item['parent'] ?? $default_id;
-			    	$this->submenu_reorder($menu_item['menu'], $this->get_new_index($menu_item, $parent), $parent);
-			    }
+					// if the item only for 'onfail' and the previous result is successful, then skip it
+					if(($menu_item['onfail'] ?? false) && $prev_action) continue;
+			    	$prev_action = $this->menu_reorder($menu_item, true);
+				}
 			}
 
 	    	if(isset($modify['rename'])) {
 			    foreach($modify['rename'] as $menu_item) {
 			    	$parent = $menu_item['parent'] ?? $default_id;
-			    	$index = $this->get_submenu_index($menu_item['menu'], $parent);
+			    	$index = $this->get_menu_index(true, $menu_item['menu'], $parent);
 			    	if($index > 0) $submenu[$parent][$index][0] = $menu_item['new_name'];
 			    }
 			}
@@ -129,7 +158,7 @@ trait zukit_AdminMenu {
 	    	if(isset($modify['remove'])) {
 			    foreach($modify['remove'] as $menu_item) {
 			    	$parent = $menu_item['parent'] ?? $default_id;
-			    	$index = $this->get_submenu_index($menu_item['menu'], $parent);
+			    	$index = $this->get_menu_index(true, $menu_item['menu'], $parent);
 			    	if($index > 0) {
 				    	unset($submenu[$parent][$index]);
 						ksort($submenu[$parent]);
@@ -138,152 +167,232 @@ trait zukit_AdminMenu {
 			}
 			// add separators if needed
 			if(isset($modify['separator'])) {
-			    foreach($modify['separator'] as $menu_item) {
+zu_log_if(self::$if_log, $modify['separator'], $this->menu_items($submenu[$default_id]));
+			    foreach($modify['separator'] as $position => $menu_item) {
+					// старый формат - массив массивов, т.е. массив для каждого элемента
+					// новый формат - массив, где ключ это позиция, а значение - привязка к элементу
+					if(is_string($menu_item)) $menu_item = [$position => $menu_item];
+					// else [, $position] = $this->key_and_pos($menu_item);
 			    	$parent = $menu_item['parent'] ?? $default_id;
-			    	$index = $this->get_new_index($menu_item, $parent);
-			    	if($index > 0 && !isset($submenu[$parent][$index])) {
-				    	$submenu[$parent][$index] = ['','read', 'separator'.$index, '', 'wp-menu-separator'];
-						ksort($submenu[$parent]);
+			    	$sep_index = $this->get_new_index($menu_item, $parent, true);
+			    	if($sep_index > 0 && $sep_index < $this->total_index_break) {
+// zu_log_if(self::$if_log, $position, $sep_index, $this->menu_items($submenu[$parent]));
+// 						$skip_separator = $this->skip_separator($submenu[$parent], $sep_index, $position);
+// zu_log_if(self::$if_log, 'skip_separator results', $skip_separator);
+// 						if($skip_separator) continue;
+						$name = sprintf('sep-%s-%s-%s', $sep_index, $position, array_values($menu_item)[0]);
+						$separator = ['','read', $name, '', 'wp-menu-separator'];
+						$update = $this->array_insert($submenu[$parent], $sep_index, $separator);
+// 						$submenu[$parent] = $update;
+						$submenu[$parent] = $this->fix_separators($update);
+zu_log_if(self::$if_log, 'after separator inserted', $this->menu_items($update));
 					}
 			    }
 			}
 		}
 
 		// output menu order for debug purpose
-		if($this->custom_menu_debug()) {
-			$this->debug_print();
+		if(--self::$menu_configs === 0 && self::$menu_debug) {
 			$this->debug_print(true);
+			$this->debug_print(true, true);
 		}
-
 	    return $menu_order;
 	}
 
-	protected function get_new_index($menu_item, $submenu_parent = '') {
-	    $new_index = $menu_item['new_index'] ?? -1;
-	    if($new_index < 0) {
-		    $base_key = array_values(array_intersect(array_keys($menu_item), [
-				'before_index',
-				'before_index2',
-				'after_index',
-				'after_index2'
-			]));
-		    $base_menu = empty($base_key) ? '' : $menu_item[$base_key[0]];
-		    $base_index = empty($submenu_parent) ?
-				$this->get_menu_index($base_menu)
-				:
-				$this->get_submenu_index($base_menu, $submenu_parent);
-		    if($base_index < 0) return (PHP_INT_MAX - 1);
-		    $index_shift = intval(filter_var($base_key[0], FILTER_SANITIZE_NUMBER_INT)) ? : 1;
-			$new_index = strpos($base_key[0], 'before') === false ? $base_index + $index_shift : $base_index - $index_shift;
-	    }
-		return $new_index;
-	}
+	/// REFACTORED /////////////////////////////////////////////////////////////
 
-	protected function get_menu_index($menu_item) {
-		global $menu;
-
-		$index = -1;
-	    foreach($menu as $key => $details) {
-	        if($details[2] == $menu_item) {
-	            $index = $key;
-	        }
-	    }
-	    return $index;
-	}
-
-	protected function menu_reorder($menu_item, $new_index) {
-		global $menu;
-
-		$index = $this->get_menu_index($menu_item);
-	    if($index > 0) {
-		    $menu[$new_index] = $menu[$index];
-		    unset($menu[$index]);
-			// Reorder the menu based on the keys in ascending order
-		    ksort($menu);
-	    }
-	}
-
-	protected function get_submenu_index($menu_item, $submenu_parent = null) {
+	protected function submenu_move($from_index, $to_index, $count = 1, $add_separator = true, $parent = null) {
 		global $submenu;
 
-		if(empty($submenu_parent)) $submenu_parent = self::$default_menu_id;
-
-		$index = -1;
-		// Get submenu key location based on slug
-	    $subitems = $submenu[$submenu_parent] ?? [];
-	    foreach($subitems as $key => $details) {
-	        if($details[2] == $menu_item) {
-	            $index = $key;
-	        }
-	    }
-	    return $index;
-	}
-
-	protected function submenu_reorder($menu_item, $new_index, $submenu_parent) {
-		global $submenu;
-
-		$index = $this->get_submenu_index($menu_item, $submenu_parent);
-	    if($index > 0) {
-		    $submenu[$submenu_parent][$new_index] = $submenu[$submenu_parent][$index];
-		    unset($submenu[$submenu_parent][$index]);
-			// Reorder the menu based on the keys in ascending order
-		    ksort($submenu[$submenu_parent]);
-	    }
-	}
-
-	protected function submenu_move($from_index, $to_index, $count = 1, $add_separator = true, $submenu_parent = null) {
-		global $submenu;
-
-		if(empty($submenu_parent)) $submenu_parent = self::$default_menu_id;
+		if(empty($parent)) $parent = self::$default_menu_id;
 
 		$first_was_separator = $count;
 		while($count > 0) {
 			if($to_index >= $this->total_index_break) break;
 
-			if(isset($submenu[$submenu_parent][$from_index])) {
-				$move_item = $submenu[$submenu_parent][$from_index];
+			if(isset($submenu[$parent][$from_index])) {
+				$move_item = $submenu[$parent][$from_index];
 				// чтобы предотвратить ситуацию когда два сепаратора идут друг за другом
-				if($first_was_separator === $count) {
-					$first_was_separator = ($move_item[4] ?? null) === 'wp-menu-separator' ? -2 : -1;
-				}
-				unset($submenu[$submenu_parent][$from_index]);
+				// if($first_was_separator === $count) {
+				// 	$first_was_separator = ($move_item[4] ?? null) === 'wp-menu-separator' ? -2 : -1;
+				// }
+				unset($submenu[$parent][$from_index]);
 				while($to_index < $this->total_index_break) {
-					if(isset($submenu[$submenu_parent][$to_index])) {
+					if(isset($submenu[$parent][$to_index])) {
 						$to_index++;
 						continue;
 					}
-					if($add_separator && $first_was_separator !== -2) {
-						$submenu[$submenu_parent][$to_index++] = ['','read', 'separator_moved_position', '', 'wp-menu-separator'];
+					if($add_separator) { // && $first_was_separator !== -2) {
+						$submenu[$parent][$to_index++] = ['','read', 'sep-moved-position', '', 'wp-menu-separator'];
 						$add_separator = false;
 						continue;
 					}
 					if(self::$menu_moved_index === null) self::$menu_moved_index = $to_index;
-					$submenu[$submenu_parent][$to_index++] = $move_item;
+					$submenu[$parent][$to_index++] = $move_item;
 					$count--;
 					break;
 				}
 			}
 			$from_index++;
 		}
-
-		// Reorder the menu based on the keys in ascending order
-	    if(isset($submenu[$submenu_parent])) ksort($submenu[$submenu_parent]);
+		// reorder the menu based on the keys in ascending order
+	    if(isset($submenu[$parent])) ksort($submenu[$parent]);
 	}
 
-	private function debug_print($is_menu = false) {
+	// get menu or submenu key location based on slug
+	protected function get_menu_index($is_submenu, $item, $parent = null) {
+		global $submenu, $menu;
+		$index = -1;
+		$items = $is_submenu ? $submenu[$parent ?: self::$default_menu_id] ?? [] : $menu;
+	    foreach($items as $key => $details) {
+	        if($details[2] === $item) return $key;
+	    }
+	    return $index;
+	}
+
+	protected function get_new_index($menu_item, $parent = null, $for_separator = false) {
+	    $new_index = $menu_item['index'] ?? $menu_item['new_index'] ?? -1;
+	    if($new_index < 0) {
+			[$item_key, $item_position] = $this->key_and_pos($menu_item);
+		    $start_index = $this->get_menu_index(!empty($parent), $item_key, $parent);
+		    if($start_index < 0) return (PHP_INT_MAX - 1);
+		    $shift = intval(filter_var($item_position, FILTER_SANITIZE_NUMBER_INT)) ?: 1;
+			$new_index = $start_index + $shift * (strpos($item_position, 'before') === false ? 1 : -1);
+			if($for_separator && strpos($item_position, 'before') !== false) $new_index = $start_index;
+	    }
+		return $new_index;
+	}
+
+	protected function menu_reorder($menu_item, $is_submenu) {
 		global $menu, $submenu;
+		$parent = $is_submenu ? $menu_item['parent'] ?? self::$default_menu_id : null;
+		$index = $this->get_menu_index($is_submenu, $menu_item['menu'], $parent);
+		$new_index = $this->get_new_index($menu_item, $parent);
 
-		$context = sprintf('*%s Order', $is_menu ? 'Menu' : 'Options Subnemu');
-		$selected = $is_menu ? $menu : $submenu[self::$default_menu_id];
+	    if($index > 0 && $new_index < $this->total_index_break && $index !== $new_index) {
+			$item = $is_submenu ? $submenu[$parent][$index] : $menu[$index];
+			if($is_submenu) unset($submenu[$parent][$index]);
+			else unset($menu[$index]);
+			$update = $this->array_insert($is_submenu ? $submenu[$parent] : $menu, $new_index, $item);
+			if($is_submenu) $submenu[$parent] = $update;
+			else $menu = $update;
 
-		$items = array_map(function($item) {
+			if(self::$detailed_log) {
+				$calledClass = static::class;
+				zu_logc("!Submenu Reorder [$calledClass]", $index, $new_index, $this->menu_items($update));
+			}
+			return true;
+	    }
+		return false;
+	}
+
+	// сложная функция которая вставляет новый элемент в массив на указанную позицию и
+	// при это старается не менять текущие индексы. Если указанное место в массиве уже
+	// занято, то тогда все элементы от этого места сдвигаются на одни вниз (и их индексы тоже)
+	private function array_insert($array, $index, $value) {
+		if(isset($array[$index])) {
+			$keys = array_keys($array);
+			$before_items = array_search($index, $keys);
+			$before = array_slice($array, 0, $before_items, true);
+			$after = array_slice($array, $before_items, null, true);
+			$akeys = array_reverse(array_keys($after));
+			foreach($akeys as $a_index) {
+				$after[$a_index + 1] = $after[$a_index];
+				unset($after[$a_index]);
+			}
+			$array = array_replace($before, $after, [$index => $value]);
+		} else {
+			$array[$index] = $value;
+		}
+		// reorder based on the keys in ascending order
+		ksort($array);
+		return $array;
+	}
+
+	private function fix_separators($menu) {
+		$prev_separator = false;
+		foreach($menu as $index => $item) {
+			$has_separator = ($item[4] ?? null) === 'wp-menu-separator';
+			if($has_separator && $prev_separator) $menu[$index] = null;
+			else $prev_separator = $has_separator;
+		}
+		return array_filter($menu);
+	}
+
+// 	private function skip_separator($menu, $index, $position) {
+// zu_log_if(self::$if_log, 'skip_separator [current]', $index, $position, $this->has_separator($menu, $index));
+// 		$has_item = $this->has_separator($menu, $index);
+// 		// if(is_null($has_item) && $position === 'before') return false;
+// 		if($has_item) return true;
+//
+// 		$keys = array_keys($menu);
+// 		if($has_item === false || (is_null($has_item) && $position === 'before')) {
+// 			$prev_index = $this->search_separator($keys, $index, true);
+// zu_log_if(self::$if_log, 'skip_separator [prev]', $keys, $prev_index, $this->has_separator($menu, $prev_index));
+// 			if($this->has_separator($menu, $prev_index)) return true;
+// 		}
+// 		if($has_item === false || (is_null($has_item) && $position === 'after')) {
+// 			$next_index = $this->search_separator($keys, $index, false);
+// zu_log_if(self::$if_log, 'skip_separator [next]', $keys, $next_index, $this->has_separator($menu, $next_index));
+// 			if($this->has_separator($menu, $next_index)) return true;
+// 		}
+// 		return false;
+// 	}
+//
+// 	private function has_separator($menu, $index) {
+// 		$item = $menu[$index] ?? [];
+// 		return ($item[4] ?? null) === 'wp-menu-separator' ? true : (count($item) ? null : false);
+// 	}
+//
+// 	private function search_separator($keys, $index, $back) {
+// 		$item_index = false;
+// 		$last_index = $keys[count($keys)-1];
+// 		while($index && $index < $last_index && !$item_index) {
+// 			$index += $back ? -1 : 1;
+// 			$item_index = array_search($index, $keys);
+// 		}
+// 		return $item_index ? $keys[$item_index] : false;
+// 	}
+
+	private function key_and_pos($item) {
+		$index = $this->snippets('array_pick_keys', $item, [
+			'before',
+			'after',
+			// устаревшие ключи
+			'before_index',
+			'before_index2',
+			'after_index',
+			'after_index2',
+		]);
+		$key = array_values($index)[0] ?? null;
+		$position = array_keys($index)[0] ?? null;
+		return [$key, preg_replace('/_index[\d]*$/', '', $position)];
+	}
+
+	private function debug_print($with_selected = false, $is_menu = false, $is_origin = false) {
+		global $menu, $submenu;
+		$context = sprintf('*%s Order', $is_origin ? 'Initial' : ($is_menu ? '"Menu"' : '"Options Subnemu"'));
+		$detailed = $is_menu ? $menu : $submenu[self::$default_menu_id];
+		$items = $this->menu_items($is_menu ? $menu : $submenu[self::$default_menu_id]);
+		if($is_origin) {
+			$submenu_items = $items;
+			$menu_items = $this->menu_items($menu);
+			zu_logc($context, $submenu_items, $menu_items);
+		} else {
+			if($with_selected && !$is_menu && !$is_origin) zu_logc('*Split Index', self::$menu_split_index);
+			if($with_selected) zu_logc($context, $items, $detailed);
+			else zu_logc($context, $items);
+		}
+	}
+
+	private function menu_items($menu) {
+		return array_map(function($item) {
 			$is_wrong = !(is_array($item) && count($item) > 2);
 			return sprintf('%s',
 				$is_wrong ? '?' : (
 					empty($item[0]) ? '-----'.$item[2].'-----' : strip_tags($item[0])
 				)
 			);
-		}, $is_menu ? $menu : $submenu[self::$default_menu_id]);
-		$this->logc($context, $items, $selected);
+		}, $menu ?? []);
 	}
 }